requirements.txt
streamlit
requests
sentence-transformers
scikit-learn
numpy
python-dotenv
